[AIRLFOW-YYY] Move variable commands to separate file
